+++ spec/minfraud_spec.rb
@@ -12,16 +12,6 @@ describe Minfraud do
       expect { |b| Minfraud.configure(&b) }.to yield_with_args(Minfraud)
     end
 
-    it 'sets user_id and license key' do
-      Minfraud.configure do |c|
-        c.user_id     = 'dummy'
-        c.license_key = 'dummy'
-      end
-
-      expect(Minfraud.user_id).to eq 'dummy'
-      expect(Minfraud.license_key).to eq 'dummy'
-    end
-
     it 'sets account_id and license key' do
       Minfraud.configure do |c|
         c.account_id  = 123
@@ -30,12 +20,6 @@ describe Minfraud do
 
       expect(Minfraud.account_id).to eq 123
       expect(Minfraud.license_key).to eq 'dummy'
-    end
-  end
-
-  describe '.configuration' do
-    it 'returns an instance of Hash' do
-      expect(Minfraud.configuration).to be_an_instance_of Hash
     end
   end
 end